+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityEditComponentPanel.java
@@ -142,8 +142,9 @@ public class EntityEditComponentPanel extends JPanel {
   private final Value<Attribute<?>> afterInsertFocusAttribute = Value.value();
 
   private final State transferFocusOnEnter = State.state(true);
-  private final Value<Integer> defaultTextFieldColumns = Value.value(DEFAULT_TEXT_FIELD_COLUMNS.get(), DEFAULT_TEXT_FIELD_COLUMNS.get());
   private final State modifiedIndicator = State.state(MODIFIED_INDICATOR.get());
+
+  private final Defaults defaults = new Defaults();
 
   /**
    * Instantiates a new EntityEditComponentPanel
@@ -348,6 +349,13 @@ public class EntityEditComponentPanel extends JPanel {
   }
 
   /**
+   * @return the {@link Defaults} instance for this edit component panel
+   */
+  protected final Defaults defaults() {
+    return defaults;
+  }
+
+  /**
    * If set to true then component labels will indicate that the value is modified.
    * This applies to all components created by this edit component panel as well as
    * components set via {@link #component(Attribute)} as long
@@ -369,18 +377,6 @@ public class EntityEditComponentPanel extends JPanel {
    */
   protected final State transferFocusOnEnter() {
     return transferFocusOnEnter;
-  }
-
-  /**
-   * Controls the default number of text field columns, -1 for not settings the columns
-   * @return the Value controlling the default number of text field columns
-   * @see #DEFAULT_TEXT_FIELD_COLUMNS
-   * @see #createTextField(Attribute)
-   * @see #createForeignKeySearchField(ForeignKey)
-   * @see #createTextFieldPanel(Attribute)
-   */
-  protected final Value<Integer> defaultTextFieldColumns() {
-    return defaultTextFieldColumns;
   }
 
   /**
@@ -533,7 +529,7 @@ public class EntityEditComponentPanel extends JPanel {
    */
   protected final TextFieldPanel.Builder createTextFieldPanel(Attribute<String> attribute) {
     return setComponentBuilder(attribute, entityComponents.textFieldPanel(attribute)
-            .columns(defaultTextFieldColumns.get())
+            .columns(defaults.textFieldColumns.get())
             .onBuild(inputPanel -> addValidator(attribute, inputPanel.textField(), editModel())));
   }
 
@@ -570,7 +566,7 @@ public class EntityEditComponentPanel extends JPanel {
    */
   protected final <T, C extends JTextField, B extends TextFieldBuilder<T, C, B>> TextFieldBuilder<T, C, B> createTextField(Attribute<T> attribute) {
     return setComponentBuilder(attribute, (TextFieldBuilder<T, C, B>) entityComponents.textField(attribute)
-            .columns(defaultTextFieldColumns.get())
+            .columns(defaults.textFieldColumns.get())
             .onBuild(field -> addFormattedValidator(attribute, field, editModel())));
   }
 
@@ -652,7 +648,7 @@ public class EntityEditComponentPanel extends JPanel {
    */
   protected final NumberField.Builder<Integer> createIntegerField(Attribute<Integer> attribute) {
     return setComponentBuilder(attribute, entityComponents.integerField(attribute)
-            .columns(defaultTextFieldColumns.get())
+            .columns(defaults.textFieldColumns.get())
             .onBuild(field -> addValidator(attribute, field, editModel())));
   }
 
@@ -663,7 +659,7 @@ public class EntityEditComponentPanel extends JPanel {
    */
   protected final NumberField.Builder<Long> createLongField(Attribute<Long> attribute) {
     return setComponentBuilder(attribute, entityComponents.longField(attribute)
-            .columns(defaultTextFieldColumns.get())
+            .columns(defaults.textFieldColumns.get())
             .onBuild(field -> addValidator(attribute, field, editModel())));
   }
 
@@ -674,7 +670,7 @@ public class EntityEditComponentPanel extends JPanel {
    */
   protected final NumberField.Builder<Double> createDoubleField(Attribute<Double> attribute) {
     return setComponentBuilder(attribute, entityComponents.doubleField(attribute)
-            .columns(defaultTextFieldColumns.get())
+            .columns(defaults.textFieldColumns.get())
             .onBuild(field -> addValidator(attribute, field, editModel())));
   }
 
@@ -685,7 +681,7 @@ public class EntityEditComponentPanel extends JPanel {
    */
   protected final NumberField.Builder<BigDecimal> createBigDecimalField(Attribute<BigDecimal> attribute) {
     return setComponentBuilder(attribute, entityComponents.bigDecimalField(attribute)
-            .columns(defaultTextFieldColumns.get())
+            .columns(defaults.textFieldColumns.get())
             .onBuild(field -> addValidator(attribute, field, editModel())));
   }
 
@@ -728,7 +724,8 @@ public class EntityEditComponentPanel extends JPanel {
    * @return a combo box builder
    */
   protected final <T, C extends JComboBox<T>, B extends ComboBoxBuilder<T, C, B>> ComboBoxBuilder<T, C, B> createComboBox(Attribute<T> attribute, ComboBoxModel<T> comboBoxModel) {
-    return setComponentBuilder(attribute, entityComponents.comboBox(attribute, comboBoxModel));
+    return (ComboBoxBuilder<T, C, B>) setComponentBuilder(attribute, entityComponents.comboBox(attribute, comboBoxModel)
+            .preferredWidth(defaults.comboBoxPreferredWidth.get()));
   }
 
   /**
@@ -738,7 +735,8 @@ public class EntityEditComponentPanel extends JPanel {
    * @return a value item list combo box builder
    */
   protected final <T> ItemComboBoxBuilder<T> createItemComboBox(Attribute<T> attribute) {
-    return setComponentBuilder(attribute, entityComponents.itemComboBox(attribute));
+    return setComponentBuilder(attribute, entityComponents.itemComboBox(attribute)
+            .preferredWidth(defaults.itemComboBoxPreferredWidth.get()));
   }
 
   /**
@@ -754,6 +752,7 @@ public class EntityEditComponentPanel extends JPanel {
     comboBoxModel.refresher().addRefreshFailedListener(this::onException);
 
     return (ComboBoxBuilder<T, C, B>) setComponentBuilder(column, entityComponents.comboBox(column, comboBoxModel)
+            .preferredWidth(defaults.comboBoxPreferredWidth.get())
             .onSetVisible(EntityEditComponentPanel::refreshIfCleared));
   }
 
@@ -768,6 +767,7 @@ public class EntityEditComponentPanel extends JPanel {
     comboBoxModel.refresher().addRefreshFailedListener(this::onException);
 
     return (ComboBoxBuilder<Entity, EntityComboBox, B>) setComponentBuilder(foreignKey, entityComponents.foreignKeyComboBox(foreignKey, comboBoxModel)
+            .preferredWidth(defaults.foreignKeyComboBoxPreferredWidth.get())
             .onSetVisible(EntityEditComponentPanel::refreshIfCleared));
   }
 
@@ -783,6 +783,7 @@ public class EntityEditComponentPanel extends JPanel {
     comboBoxModel.refresher().addRefreshFailedListener(this::onException);
 
     return setComponentBuilder(foreignKey, entityComponents.foreignKeyComboBoxPanel(foreignKey, comboBoxModel, editPanelSupplier))
+            .comboBoxPreferredWidth(defaults.foreignKeyComboBoxPreferredWidth.get())
             .onSetVisible(entityComboBoxPanel -> refreshIfCleared(entityComboBoxPanel.comboBox()));
   }
 
@@ -794,7 +795,7 @@ public class EntityEditComponentPanel extends JPanel {
   protected final EntitySearchField.Builder createForeignKeySearchField(ForeignKey foreignKey) {
     return setComponentBuilder(foreignKey, entityComponents.foreignKeySearchField(foreignKey,
                     editModel().foreignKeySearchModel(foreignKey))
-            .columns(defaultTextFieldColumns.get()));
+            .columns(defaults.foreignKeySearchFieldColumns.get()));
   }
 
   /**
@@ -807,7 +808,7 @@ public class EntityEditComponentPanel extends JPanel {
                                                                                   Supplier<EntityEditPanel> editPanelSupplier) {
     return setComponentBuilder(foreignKey, entityComponents.foreignKeySearchFieldPanel(foreignKey,
                     editModel().foreignKeySearchModel(foreignKey), editPanelSupplier)
-            .columns(defaultTextFieldColumns.get()));
+            .columns(defaults.foreignKeySearchFieldColumns.get()));
   }
 
   /**
@@ -975,6 +976,64 @@ public class EntityEditComponentPanel extends JPanel {
     return Arrays.stream(parent.getComponents()).anyMatch(childComponent ->
             childComponent instanceof JComponent &&
                     sameOrParentOf((JComponent) childComponent, component));
+  }
+
+  /**
+   * Specifies the availible default values for component builders.
+   */
+  protected static final class Defaults {
+
+    private final Value<Integer> textFieldColumns = Value.value(DEFAULT_TEXT_FIELD_COLUMNS.get(), DEFAULT_TEXT_FIELD_COLUMNS.get());
+    private final Value<Integer> foreignKeySearchFieldColumns = Value.value(DEFAULT_TEXT_FIELD_COLUMNS.get(), DEFAULT_TEXT_FIELD_COLUMNS.get());
+    private final Value<Integer> foreignKeyComboBoxPreferredWidth = Value.value(0, 0);
+    private final Value<Integer> itemComboBoxPreferredWidth = Value.value(0, 0);
+    private final Value<Integer> comboBoxPreferredWidth = Value.value(0, 0);
+
+    /**
+     * Controls the default number of text field columns, -1 for not settings the columns
+     * @return the Value controlling the default number of text field columns
+     * @see #DEFAULT_TEXT_FIELD_COLUMNS
+     * @see #createTextField(Attribute)
+     * @see #createTextFieldPanel(Attribute)
+     */
+    public Value<Integer> textFieldColumns() {
+      return textFieldColumns;
+    }
+
+    /**
+     * Controls the default number of text field columns, -1 for not settings the columns
+     * @return the Value controlling the default number of foreign key search field columns
+     * @see #DEFAULT_TEXT_FIELD_COLUMNS
+     * @see #createForeignKeySearchField(ForeignKey)
+     */
+    public Value<Integer> foreignKeySearchFieldColumns() {
+      return foreignKeySearchFieldColumns;
+    }
+
+    /**
+     * @return the Value controlling the default combo box width
+     * @see #createComboBox(Column)
+     */
+    public Value<Integer> comboBoxPreferredWidth() {
+      return comboBoxPreferredWidth;
+    }
+
+    /**
+     * @return the Value controlling the default item combo box width
+     * @see #createItemComboBox(Attribute)
+     */
+    public Value<Integer> itemComboBoxPreferredWidth() {
+      return itemComboBoxPreferredWidth;
+    }
+
+    /**
+     * @return the Value controlling the default foreign key combo box width
+     * @see #createForeignKeyComboBox(ForeignKey)
+     * @see #createForeignKeyComboBoxPanel(ForeignKey, Supplier)
+     */
+    public Value<Integer> foreignKeyComboBoxPreferredWidth() {
+      return foreignKeyComboBoxPreferredWidth;
+    }
   }
 
   private final class OnComponentBuilt<C extends JComponent> implements Consumer<C> {