+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntityComboBoxPanel.java
@@ -30,6 +30,7 @@ import is.codion.swing.common.ui.key.TransferFocusOnEnter;
 import is.codion.swing.framework.model.component.EntityComboBoxModel;
 import is.codion.swing.framework.ui.EntityEditPanel;
 
+import javax.swing.AbstractButton;
 import javax.swing.Action;
 import javax.swing.JPanel;
 import javax.swing.KeyStroke;
@@ -73,6 +74,7 @@ public final class EntityComboBoxPanel extends JPanel {
   }
 
   private final EntityComboBox comboBox;
+  private final List<AbstractButton> buttons = new ArrayList<>(0);
 
   private EntityComboBoxPanel(DefaultBuilder builder) {
     comboBox = builder.createComboBox();
@@ -87,7 +89,7 @@ public final class EntityComboBoxPanel extends JPanel {
     }
     setLayout(new BorderLayout());
     add(createButtonPanel(comboBox, builder.buttonsFocusable, builder.buttonLocation,
-            actions.toArray(new Action[0])), BorderLayout.CENTER);
+            buttons, actions.toArray(new Action[0])), BorderLayout.CENTER);
     addFocusListener(new InputFocusAdapter(comboBox));
   }
 
@@ -158,6 +160,12 @@ public final class EntityComboBoxPanel extends JPanel {
      * @return this builder instance
      */
     Builder keyStroke(KeyboardShortcut keyboardShortcut, KeyStroke keyStroke);
+
+    /**
+     * @param comboBoxPreferredWidth the preferred combo box width
+     * @return this builder instance
+     */
+    Builder comboBoxPreferredWidth(int comboBoxPreferredWidth);
 
     /**
      * @return a new {@link EntityComboBoxPanel} based on this builder
@@ -235,6 +243,12 @@ public final class EntityComboBoxPanel extends JPanel {
     }
 
     @Override
+    public Builder comboBoxPreferredWidth(int comboBoxPreferredWidth) {
+      entityComboBoxBuilder.preferredWidth(comboBoxPreferredWidth);
+      return this;
+    }
+
+    @Override
     protected EntityComboBoxPanel createComponent() {
       return new EntityComboBoxPanel(this);
     }
@@ -247,6 +261,7 @@ public final class EntityComboBoxPanel extends JPanel {
     @Override
     protected void enableTransferFocusOnEnter(EntityComboBoxPanel component) {
       TransferFocusOnEnter.enable(component.comboBox());
+      component.buttons.forEach(TransferFocusOnEnter::enable);
     }
 
     @Override