+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntityControls.java
@@ -36,6 +36,7 @@ import is.codion.swing.framework.model.component.EntityComboBoxModel;
 import is.codion.swing.framework.ui.EntityEditPanel;
 import is.codion.swing.framework.ui.icon.FrameworkIcons;
 
+import javax.swing.AbstractButton;
 import javax.swing.Action;
 import javax.swing.JComponent;
 import javax.swing.JDialog;
@@ -127,7 +128,8 @@ final class EntityControls {
   }
 
   static JPanel createButtonPanel(JComponent centerComponent, boolean buttonFocusable,
-                                  String borderLayoutConstraints, Action... buttonActions) {
+                                  String borderLayoutConstraints, List<AbstractButton> buttons,
+                                  Action... buttonActions) {
     Dimension preferredSize = centerComponent.getPreferredSize();
 
     return Components.panel(new BorderLayout())
@@ -136,6 +138,7 @@ final class EntityControls {
                     .buttonsFocusable(buttonFocusable)
                     .preferredButtonSize(new Dimension(preferredSize.height, preferredSize.height))
                     .buttonGap(0)
+                    .buttonBuilder(buttonBuilder -> buttonBuilder.onBuild(buttons::add))
                     .build(), borderLayoutConstraints)
             .build();
   }